+++ tests/test_ecephys/test_tools_spikeextractors.py
@@ -171,21 +171,21 @@ class TestExtractors(unittest.TestCase):
             nwbfile_path=path_multi,
             metadata=metadata,
             write_as="raw",
-            es_key="ElectricalSeries_raw",
+            es_key="ElectricalSeriesRaw",
         )
         write_recording(
             recording=self.RX2,
             nwbfile_path=path_multi,
             metadata=metadata,
             write_as="processed",
-            es_key="ElectricalSeries_processed",
+            es_key="ElectricalSeriesProcessed",
         )
         write_recording(
             recording=self.RX3,
             nwbfile_path=path_multi,
             metadata=metadata,
             write_as="lfp",
-            es_key="ElectricalSeries_lfp",
+            es_key="ElectricalSeriesLfp",
         )
 
         RX_nwb = se.NwbRecordingExtractor(file_path=path_multi, electrical_series_name="raw_traces")
@@ -203,7 +203,7 @@ class TestExtractors(unittest.TestCase):
         compression = "gzip"
         with NWBHDF5IO(path=path, mode="r") as io:
             nwbfile = io.read()
-            compression_out = nwbfile.acquisition["ElectricalSeries_raw"].data.compression
+            compression_out = nwbfile.acquisition["ElectricalSeriesRaw"].data.compression
         self.assertEqual(
             compression_out,
             compression,
@@ -221,7 +221,7 @@ class TestExtractors(unittest.TestCase):
         )
         with NWBHDF5IO(path=path, mode="r") as io:
             nwbfile = io.read()
-            compression_out = nwbfile.acquisition["ElectricalSeries_raw"].data.compression
+            compression_out = nwbfile.acquisition["ElectricalSeriesRaw"].data.compression
         self.assertEqual(
             compression_out,
             compression,
@@ -240,7 +240,7 @@ class TestExtractors(unittest.TestCase):
         )
         with NWBHDF5IO(path=path, mode="r") as io:
             nwbfile = io.read()
-            compression_out = nwbfile.acquisition["ElectricalSeries_raw"].data.compression
+            compression_out = nwbfile.acquisition["ElectricalSeriesRaw"].data.compression
         self.assertEqual(
             compression_out,
             compression,
@@ -259,7 +259,7 @@ class TestExtractors(unittest.TestCase):
         )
         with NWBHDF5IO(path=path, mode="r") as io:
             nwbfile = io.read()
-            compression_out = nwbfile.acquisition["ElectricalSeries_raw"].data.compression
+            compression_out = nwbfile.acquisition["ElectricalSeriesRaw"].data.compression
         self.assertEqual(
             compression_out,
             compression,
@@ -274,7 +274,7 @@ class TestExtractors(unittest.TestCase):
         write_recording(recording=self.RX, nwbfile_path=path, overwrite=True, metadata=self.placeholder_metadata)
         with NWBHDF5IO(path=path, mode="r") as io:
             nwbfile = io.read()
-            chunks_out = nwbfile.acquisition["ElectricalSeries_raw"].data.chunks
+            chunks_out = nwbfile.acquisition["ElectricalSeriesRaw"].data.chunks
         test_iterator = SpikeInterfaceRecordingDataChunkIterator(recording=self.RX)
         self.assertEqual(
             chunks_out,